+++ CHANGELOG.md
@@ -25,6 +25,7 @@ FlatLaf Change Log
   `TitlePane.unifiedBackground` is `true`.
 - Button and ToggleButton: Do not paint background of disabled (and unselected)
   toolBar buttons. (issue #292; regression since fixing #112)
+- SplitPane: `JSplitPane.setContinuousLayout(false)` did not work. (issue #301)
 - TabbedPane: Fixed NPE when creating/modifying in another thread. (issue #299)
 - Fixed crash when running in Webswing. (issue #290)